refactored types

+++ src/structures.h
@@ -2,17 +2,17 @@
 #define STRUCTURES_H
 
 struct Adjacency {
-  int ** matrix;
+  short int ** matrix;
   size_t dim;
 } * A1, * A2;
 
 struct Graph {
   short int * taken;
   short int * translate;
-  int ** edges;
+  short int ** edges;
   short int * num_outgoing;
-  short int num_nodes;
-  int num_edges;
+  unsigned int num_nodes;
+  unsigned int num_edges;
   char ** id2name;
 } * G1, * G2;
